// web/src/api/generated/model/claimDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ClaimLineDto } from './claimLineDto';
import type { ClaimStatus } from './claimStatus';

export interface ClaimDto {
  id: string;
  number: string;
  employeeId: string;
  date: string;
  /** @nullable */
  memo: string | null;
  status: ClaimStatus;
  total: number;
  /** @nullable */
  rejectionReason: string | null;
  /** @nullable */
  voucherId: string | null;
  /** @nullable */
  paymentVoucherId: string | null;
  /** @nullable */
  paidDate: string | null;
  lines: ClaimLineDto[];
}