// web/src/api/generated/model/claimInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ClaimLineInput } from './claimLineInput';

export interface ClaimInput {
  employeeId: string;
  date: string;
  /** @nullable */
  memo: string | null;
  lines: ClaimLineInput[];
}