// web/src/api/generated/model/claimStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type ClaimStatus = typeof ClaimStatus[keyof typeof ClaimStatus];


export const ClaimStatus = {
  Draft: 'Draft',
  Submitted: 'Submitted',
  Approved: 'Approved',
  Rejected: 'Rejected',
  Paid: 'Paid',
} as const;